Fix: ToB-21 (unsafe casts) (#1457)
* Add ChainContext library

* Add coverage for the new library

* Remove unsafe casts for getting chain context

* Safe cast for attestation nonces

* Make `stateIndex` uint8

* Adjust tests with uint8 stateIndex

* Add safe casts for 32 bits indexes

* Document dispute index unsafe cast

* Document unsafe casts in `memory` libs

* Document unsafe casts which are actually upcasts

* Document remaining unsafe casts

* Chore: better formatting

* Add SafeCast for `encodeTips256()`

* Bound gasDrop/gasLimit max values to avoid overflow in tests

* Revert changes that are already in #1430

* Better docs for snapshot merkle root

* Chore: explicit reverts when bounding in tests

* Add tests with potential context overflows

+++ packages/contracts-core/contracts/interfaces/ISnapshotHub.sol
@@ -93,5 +93,5 @@ interface ISnapshotHub {
      * @param stateIndex    Index of state in the attestation's snapshot
      * @return snapProof    The snapshot proof
      */
-    function getSnapshotProof(uint32 attNonce, uint256 stateIndex) external view returns (bytes32[] memory snapProof);
+    function getSnapshotProof(uint32 attNonce, uint8 stateIndex) external view returns (bytes32[] memory snapProof);
 }
